refactor: split federation_2d.tsx into 9 modules with Zustand store
789 LOC monolith -> types, colors, simulation, store (Zustand),
draw helpers, Canvas2D component, Sidebar component, data hook.
All files <200 LOC. Zero new dependencies.
