+++ src/components/federation/index.ts
@@ -1,0 +1,7 @@
+export { Canvas2D } from "./Canvas2D";
+export { Sidebar } from "./Sidebar";
+export { useFederationStore } from "./store";
+export { simulate } from "./simulation";
+export { machineColor, statusGlow, hexRgb } from "./colors";
+export { drawGrid, drawClusterLabels, drawEdges, drawAgents, drawLegend } from "./draw";
+export type { AgentNode, AgentEdge, Particle } from "./types";
